### 2024-07-18 #7002 @editorialbot
[PRE REVIEW]: biopixR: Extracting Insights from Biological Images
**Submitting author:** @Brauckhoff (<a href="http://orcid.org/0009-0002-0142-7017">Tim Brauckhoff</a>)
**Repository:** https://github.com/Brauckhoff/biopixR
**Branch with paper.md** (empty if default branch):
**Version:** 1.1.0
