### 2024-07-18 a86be72
added DOIs in JOSS
