+++ joss/paper.bib
@@ -153,6 +153,7 @@
   author = {Jeroen Ooms},
   note   = {R package version 2.8.3},
   year   = {2024},
+  doi    = {10.32614/CRAN.package.magick},
   url    = {https://docs.ropensci.org/magick/
 https://ropensci.r-universe.dev/magick},
 }
