### 2024-08-05 #7002 @fabian-s
:wave:  @ColemanRHarris & @david-barnett

would you be willing to review this submission for JOSS?
