### 2024-08-05 #7002 @ColemanRHarris
Hi @fabian-s, thanks for reaching out. I am willing to review the submission for JOSS! What do you need from me?
