### 2024-08-06 #7002 @david-barnett
Hi @fabian-s - I think for me the topic isn't a good fit with my expertise, I don't work with images. So I'll have to decline. Good luck with the review!
