### 2024-08-06 #7002 @fabian-s
:wave: @askerdb, @tijeco

would any of you be willing to review this submission for JOSS?
