### 2024-08-06 #7074 @editorialbot
[REVIEW]: biopixR: Extracting Insights from Biological Images
**Submitting author:** @Brauckhoff (<a href="http://orcid.org/0009-0002-0142-7017">Tim Brauckhoff</a>)
**Repository:** https://github.com/Brauckhoff/biopixR
**Branch with paper.md** (empty if default branch):
**Version:** 1.2.0
