### 2024-08-06 #7074 @fabian-s
@Brauckhoff
 while we wait for reviewers to start working on this, please start fixing the missing DOIs the bot pointed out above.
