+++ joss/paper.bib
@@ -166,6 +166,7 @@ https://ropensci.r-universe.dev/magick},
   year      = {2019},
   address   = {Boca Raton},
   isbn      = {9781138359420},
+  doi       = {10.1201/9781138359444},
   note      = {Includes bibliographical references and index},
   series    = {Chapman & Hall/CRC, the {R} series},
   pagetotal = {303},
