### 2024-08-06 #7074 @devSJR
Dear @fabian-s, thanks for pointing out the issue with the DOIs. We are eager to cite everything we use or refer to with state-of-the-art technologies. Therefore, we fixed most missing DOIs shortly after submission during the [PRE REVIEW](https://github.com/openjournals/joss-reviews/issues/7002#issuecomment-2236226396). It appears, that we missed a DOI for RMarkown, which I just added.

However, We fear, there is little we can do for the others, since these items have no DOI. For example:

- [foodwebr](https://github.com/lewinfox/foodwebr) is on GitHub only and
- [radiomics is archived on CRAN](https://cran.r-project.org/web/packages/radiomics/index.html). We cite radiomics because it is an R package, which allows extracting Haralick features and clusters information using Partitioning Around Medoids (PAM). However, we will remove the reference to the package since the reader is still guided to appropriate literature.

Is this acceptable, or do you see another solution to this?
