### 2024-08-06 2b690ae
working on a new function
