### 2024-08-07 #7074 @fabian-s
> However, we fear, there is little we can do for the others, since these items have no DOI. For example:
>
>     * [foodwebr](https://github.com/lewinfox/foodwebr) is on GitHub only and
>
>     * [radiomics is archived on CRAN](https://cran.r-project.org/web/packages/radiomics/index.html). We cite radiomics because it is an R package, which allows extracting Haralick features and clusters information using Partitioning Around Medoids (PAM). However, we will remove the reference to the package since the reader is still guided to appropriate literature.
>
> Is this acceptable, or do you see another solution to this?

@devSJR
thanks, perfect!
