### 2024-08-07 #7074 @devSJR
Just a small note. We have added a new function to the package, which didn't end up in a branch but in main. Sorry for that, we will fix that. The package submitted for review is the stable version 1.x after bug fixes (→ 1.1.0).
