### 2024-08-15 #7074 @devSJR
Dear @fabian-s, @tijeco and @ColemanRHarris,

we forgot to mention that we already fixed our latest commit, which caused a problem:

> Just a short note. We have added a new function to the package, which didn't end up in a branch but in main. Sorry for that, we will fix that. The package submitted for review is the stable version 1.x after bug fixes (→ 1.1.0).

For the moment being, we decided to halt all pending commits of experimental new functions and new data sets, since we haven't tested them intensively in our current research. We decided it is also better not to submit anything until the review process is finished. Better than saying, “Let's commit. What could possibly go wrong?”, in my humble opinion. I hope you consent with this approach.
