### 2024-08-16 #7074 @tijeco
@fabian-s , for clarity, we should be reviewing https://github.com/Brauckhoff/biopixR/tree/1880095c8bca6701b5363857f1cf4bba9738379b, corresponding to their release 1.1.0?
