### 2024-08-16 #7074 @tijeco
@devSJR  Could you provide clarification about contribution and authorship? The authors in the manuscript are Tim Brauckhoff, Coline Kieffer, and Stefan Rödiger. The first and last author's appear to have made contributions in the commit history. I don't see the middle author's contribution. Also, it looks like Michael Chirico contributed to the software via pull request, but I don't see them in the manuscript.
