### 2024-08-16 #7074 @devSJR
Dear @tijeco, sure, I can do that.

Tim does the main programming and documentation of biopixR.

I lead this project (active project direction) and usually review the code (sometimes we even did paired programming), make contributions, fixes, and help with the documentation.

Coline Kieffer contributes the various datasets that are central to the package. For this, she prepared wet lab samples, did the imaging and subsequent insight into the data jointly with us. So to speak, she provided the oil. She also, raised issues (albeit personally in the lab or by decentralized digital messaging) and reflects art of the target audience (biologists in the wet lab). We also specifically mention her in the supplement to this paper https://zenodo.org/doi/10.5281/zenodo.12744222 on page 78 (this supplement is referenced as Brauckhoff, T., & Rödiger, S. (2024). biopixR: Extracting insights from biological images. https://doi.org/10.5281/ZENODO.12744222). You will certainly notice that there are more people mentioned. However, considering invested time and dedication, Coline certainly contributed significantly.

Michael Chirico contributed, indeed a software pull request. But truth to be told, I don't know if we should/must include him. @fabian-s, would you be so kind and tell me how to deal with this?

I hope this answers your question.
