### 2024-08-19 #7074 @fabian-s
@tijeco
>  for clarity, we should be reviewing https://github.com/Brauckhoff/biopixR/tree/1880095c8bca6701b5363857f1cf4bba9738379b, corresponding to their release 1.1.0?

that's how i understood @devSJR 's remark, too.
