### 2024-08-19 #7074 @fabian-s
@devSJR

> Michael Chirico contributed, indeed a software pull request. But truth to be told, I don't know if we should/must include him. @fabian-s, would you be so kind and tell me how to deal with this?

I looked at their 2 commits, seems like very small housekeeping stuff to me. I don't think their contribs warrant a co-authorship.
