### 2024-09-30 #7074 @tijeco
Hi! I'm working on wrapping things up. I had some issues getting things to install on my mac without root priveleges. So I did get it working in a github codespace. I noticed the specific dependencies are noted in the manuscript, but they don't seem to be noted in the README, things like the version of R, imager, magick etc. I can raise that as an issue in the repo if you'd like!
