### 2024-10-01 #7074 @devSJR
@tijeco thanks for testing!

We have no Macs, but we use CI and CRAN, which gave us no issues. Thus, we missed that.

An issue would be good. These things are fixable timely.
