### 2024-10-01 #7074 @fabian-s
> @fabian-s and @devSJR Hi! I'm working on wrapping things up. I had some issues getting things to install on my mac without root priveleges. So I did get it working in a github codespace. I noticed the specific dependencies are noted in the manuscript, but they don't seem to be noted in the README, things like the version of R, imager, magick etc. I can raise that as an issue in the repo if you'd like!

![image](https://github.com/user-attachments/assets/b235cb49-6d27-4bc5-b231-f8677a2d850a)
system reqs for an R package are properly placed in `DESCRIPTION`, that's what it's for.
