### 2024-10-01 #7074 @devSJR
@fabian-s Our suggestion is to add a small Trouble Shooting section to the README. Though we pass all tests, we saw that @tijeco had an issue he could get fixed. From experience, R is cross-platform, but the devil is in the details.

@tijeco, if you were so kind and add some technical details (macOS version …) we will add a paragraph like: “biopixR was tested on many platforms and should be functional on all operating systems. However, we have received reports in which users initially had problems with the installation. Under macOS [DETAILS], please check if the following requirements are met …”
