### 2024-10-01 #7074 @tijeco
Ah okay I see it's in the DESCRIPTION file. Perhaps just a small note in the README similar to in the manuscript:
>Therefore, biopixR depends on R (≥ 4.2.0), imager, magick and tcltk, imports data.table
> and cluster

What you have in the github actions is right, and it does work on mac. I was trying to get it on a mac that I don't have sudo access to. I was hoping I could put everything in a conda environment, but imager isn't available through conda. I also couldn't install X11 with brew without sudo. I went through a long series of needlessly painful steps all because I didn't have administrator rights. Your package installs perfectly fine in a system that has the ability to install all the necessary dependencies.
