### 2024-10-01 #7074 @devSJR
@tijeco I am genuinely sorry but also thankful for your experience. @Brauckhoff got access to a Mac and is now adding this information to the README.
