### 2024-10-01 #7074 @Brauckhoff
Hello!

@devSJR I tested the installation on a Mac and found that while the `biopixR` package installs without issues, loading it can result in an error if **X11** is missing. This error occurs due to the **imager** dependency, which requires **X11** to function properly. Without it, the package won't load. Installing [**XQuartz**](https://www.xquartz.org/) resolves the issue, and after doing so, everything worked smoothly for me. I tested this on a MacBook running macOS Ventura 13.0 with R version 4.4.1. I added some information to the README to introduce the **X11** issue and noted the other dependencies, as suggested by @tijeco.
