+++ README.md
@@ -13,15 +13,15 @@
 ## Introduction
 In the rapidly evolving landscape of scientific research and technology, the field of image analysis and processing has become indispensable, providing unprecedented insight into complex phenomena. This project aims to utilize image processing techniques, using the [`R` statistical programming language](https://www.r-project.org/), to analyze and characterize bead microparticles. Microparticles find extensive utility in various scientific domains, including medical diagnostics, environmental monitoring, and materials science.
 
-The rise of high-resolution imaging technologies has led to an exponential increase in the volume and complexity of image data, requiring sophisticated computational methods for efficient analysis. In this context, `R` is a standout programming language that excels in statistical computing and graphics. Leveraging the rich ecosystem of R packages, this project aims to develop an automated image analysis pipeline for bead microparticles. 
+The rise of high-resolution imaging technologies has led to an exponential increase in the volume and complexity of image data, requiring sophisticated computational methods for efficient analysis. In this context, `R` is a standout programming language that excels in statistical computing and graphics. Leveraging the rich ecosystem of `R` packages, this project aims to develop an automated image analysis pipeline for bead microparticles. 
 
 **The primary objectives of this project include**:
  - *Image Preprocessing*: Implementing techniques to fill and reconnect discontinuous lines and edges. Discard coagulated duplets, multiplets, and beads that are too close together as they may excite each other and produce a false positive signal.
  - *Segmentation and Feature Extraction*: Applying algorithms to accurately identify and segment sperical objects (e.g, microparticles). Extracting relevant features such as quantity, size, and intensity for comprehensive characterization. 
- - *Visualization*: Implementing interactive tools with [Tcl/Tk](https://www.tcl-lang.org/) via the tcltk package for the selection of thresholds and smoothing factors. Generating images to control the appropriate functioning of algorithms, for example the reconnection of lines.
+ - *Visualization*: Implementing interactive tools with [Tcl/Tk](https://www.tcl-lang.org/) via the `tcltk` package for the selection of thresholds and smoothing factors. Generating images to control the appropriate functioning of algorithms, for example the reconnection of lines.
  - *Automatization*: Combining algorithms for medium-throughput analysis of image data.
 
-This project aims to meet the immediate need for effective bead microparticle analysis and contribute to the broader field of image processing methodologies in the R programming environment. By providing a comprehensive and adaptable framework, the work empowers researchers and practitioners to extract meaningful insights from image data, thus enhancing our understanding of bead microparticles and their diverse applications.
+This project aims to meet the immediate need for effective bead microparticle analysis and contribute to the broader field of image processing methodologies in the `R` programming environment. By providing a comprehensive and adaptable framework, the work empowers researchers and practitioners to extract meaningful insights from image data, thus enhancing our understanding of bead microparticles and their diverse applications.
 
 
 ## Installation
@@ -32,7 +32,7 @@ Now available on CRAN so try:
 install.packages("biopixR")
 ```
 
-or try the latest version of biopixR:
+or try the latest version of `biopixR`:
 
 ```{r}
 install.packages("devtools")
@@ -41,7 +41,7 @@ devtools::install_github("Brauckhoff/biopixR")
 
[…]
