### 2024-10-01 #7074 @ColemanRHarris
@Brauckhoff I ran into the same issues as @tijeco since I am on an M2 Mac. I will add my thoughts here as I go through my review, but one note on the documentation you added to the README.

The command you listed, `brew cask install xquartz`, is outdated. When I ran this, I received the following error:
```
Error: `brew cask` is no longer a `brew` command. Use `brew <command> --cask` instead.
```

Hence, the proper command should be:
```
brew install xquartz --cask
```
