### 2024-10-01 #7074 @tijeco
@ColemanRHarris , that's actually what they have in their github actions! https://github.com/Brauckhoff/biopixR/blob/1880095c8bca6701b5363857f1cf4bba9738379b/.github/workflows/R-CMD-check.yml#L58
