### 2024-10-01 #7074 @Brauckhoff
@ColemanRHarris, thank you for the correction. I rechecked the documentation in the CI, as mentioned by @tijeco, and I found a minor mistake in the order of commands that could be responsible for the error. The correct command should be `brew install --cask xquartz`, I will update this in the README.
