### 2024-10-01 #7074 @ColemanRHarris
@devSJR Updating the README is sufficient, as that will be useful for any future users with a Mac.
