### 2024-10-01 #7074 @ColemanRHarris
Only two notes on the documentation (not sure if this is for @Brauckhoff or @devSJR), otherwise everything looks and works as expected 👍

1. When loading the `biopixR` library from R at the command line, I get the following warning. This did not prevent me from running any of the documentation examples, but might be worth investigating this or suppressing the warning.

  ```
  Loading required package: tcltk
  Warning message:
  In fun(libname, pkgname) :
    no display name and no $DISPLAY environment variable
  ```

2. Change links to vignettes in the README and elsewhere from https://github.com/Brauckhoff/biopixR/blob/main/vignettes/biopixR.Rmd to https://cran.r-project.org/web/packages/biopixR/vignettes/biopixR.html so that the vignette examples are readable.
