### 2024-10-02 #7074 @Brauckhoff
@ColemanRHarris, I updated the URL in the README file as you noted.

Regarding the warning you mentioned, I tried to reproduce it on my console, but I couldn’t replicate the issue. It seems I am still able to access a display environment, even when running R through the console.

To address the warning message, I removed `tcltk` from the dependencies and moved it to the suggests section. Additionally, I added code to check for a display environment (X11) to ensure that `tcltk` is only loaded if the requirements are met. This way, we should be able to maintain the same functionality as before.
