### 2024-10-02 #7074 @ColemanRHarris
Hi @Brauckhoff, I uninstalled the package and re-installed the current version from GitHub. I do see some of the changes you made, but still get this output when loading the package, e.g. `library(biopixR)`:
```
Loading required package: magick
Linking to ImageMagick 6.9.12.93
Enabled features: cairo, fontconfig, freetype, heic, lcms, pango, raw, rsvg, webp
Disabled features: fftw, ghostscript, x11
X11 not available, skipping tcltk.
Warning message:
In fun(libname, pkgname) :
  no display name and no $DISPLAY environment variable
```

I looked through your repository, it seems like you [still import](https://github.com/Brauckhoff/biopixR/blob/0b6e2087288a4bedae1a27b72c461942818f53b3/NAMESPACE#L22) `tcltk` in the `NAMESPACE` file, have you re-run the build with your current code changes?
