+++ R/interactive_objectDetection.R
@@ -35,7 +35,6 @@ print_with_timestamp <- function(msg) {
 #' @import magick
 #' @import imager
 #' @import data.table
-#' @import tcltk
 #' @references https://CRAN.R-project.org/package=magickGUI
 #' @examples
 #' \donttest{
@@ -48,6 +47,12 @@ interactive_objectDetection <-
   function(img,
            resolution = 0.1,
            return_param = FALSE) {
+
+    # Check for required namespace
+    if (!requireNamespace("tcltk", quietly = TRUE)) {
+      stop("The 'tcltk' package is required but not installed. Please install it using install.packages('tcltk').")
+    }
+
     # Convert the image to the desired format
     image_original <- cimg2magick(img)
     object_img <- img
@@ -83,7 +88,7 @@ interactive_objectDetection <-
     temp <- tempfile(fileext = ".jpg")
     on.exit(unlink(temp), add = TRUE)
     image_write(initial, temp)
-    image_tcl <- tkimage.create("photo", "image_tcl", file = temp)
+    image_tcl <- tcltk::tkimage.create("photo", "image_tcl", file = temp)
 
     # Format label digits based on resolution
     label_digits <- -as.integer(log(resolution, 10))
@@ -91,30 +96,30 @@ interactive_objectDetection <-
     label_template <- sprintf("%%.%df", label_digits)
 
     # Create GUI window and frames
-    win1 <- tktoplevel()
-    on.exit(tkdestroy(win1), add = TRUE)
-    win1.frame1 <- tkframe(win1)
-    win1.frame2 <- tkframe(win1)
-    win1.frame3 <- tkframe(win1)
-    win1.frame4 <- tkframe(win1)
-    win1.im <- tklabel(win1, image = image_tcl)
+    win1 <- tcltk::tktoplevel()
+    on.exit(tcltk::tkdestroy(win1), add = TRUE)
+    win1.frame1 <- tcltk::tkframe(win1)
+    win1.frame2 <- tcltk::tkframe(win1)
+    win1.frame3 <- tcltk::tkframe(win1)
+    win1.frame4 <- tcltk::tkframe(win1)
+    win1.im <- tcltk::tklabel(win1, image = image_tcl)
 
     # Create and configure labels
     win1.frame1.label <-
-      tklabel(win1.frame1, text = sprintf("%s%s", text_label_alpha, sprintf(label_template, alpha)))
+      tcltk::tklabel(win1.frame1, text = sprintf("%s%s", text_label_alpha, sprintf(label_template, alpha)))
     win1.frame2.label <-
-      tklabel(win1.frame2, text = sprintf("%s%s", text_label_sigma, sprintf(label_template, sigma)))
+      tcltk::tklabel(win1.frame2, text = sprintf("%s%s", text_label_sigma, sprintf(label_template, sigma)))
     win1.frame3.label <-
-      tklabel(win1.frame3, text = sprintf("%s%s", text_label_scale, sprintf(label_template, scale)))
+      tcltk::tklabel(win1.frame3, text = sprintf("%s%s", text_label_scale, sprintf(label_template, scale)))
     win1.frame4.label <-
-      tklabel(win1.frame4, text = paste("Method: ", method))
+      tcltk::tklabel(win1.frame4, text = paste("Method: ", method))
     win1.method_status <-
-      tklabel(win1, text = paste("Current method in use: ", method))
+      tcltk::tklabel(win1, text = paste("Current method in use: ", method))
 
     # Initialize slider values
-    slider_value_alpha <- tclVar(alpha)
-    slider_value_sigma <- tclVar(sigma)
-    slider_value_scale <- tclVar(scale)
[…]
