+++ NAMESPACE
@@ -19,7 +19,6 @@ import(data.table)
 import(imager)
 import(magick)
 import(parallel)
-import(tcltk)
 importFrom(cluster,pam)
 importFrom(cluster,silhouette)
 importFrom(grDevices,col2rgb)
