+++ DESCRIPTION
@@ -37,7 +37,7 @@ Suggests:
     foreach
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.2.3
+RoxygenNote: 7.3.2
 LazyData: true
 LazyLoad: yes
 NeedsCompilation: no
