### 2024-10-02 #7074 @ColemanRHarris
@Brauckhoff Fantastic, thank you! Can confirm that I installed the updated version and produced no warnings! 👍  See output below:
```
Loading required package: magick
Linking to ImageMagick 6.9.12.93
Enabled features: cairo, fontconfig, freetype, heic, lcms, pango, raw, rsvg, webp
Disabled features: fftw, ghostscript, x11
X11 not available, skipping tcltk.
```
