### 2024-10-08 #7074 @devSJR
I have checked the proof and think it is correct.
