### 2024-10-09 #7074 @devSJR
@fabian-s is it our task to move forward with the command `@editorialbot accept`?
