### 2024-10-09 #7074 @fabian-s
nah, [@bcm-eics](https://github.com/orgs/openjournals/teams/bcm-eics) need to look this over and give their thumbs up
